requirements.txt
# Core parsing
pymupdf==1.24.10
python-docx==1.1.2
pdfplumber==0.11.4

# NLP (optional)
spacy==3.7.4

# Data handling
pandas==2.2.2
openpyxl==3.1.5

# Web API
fastapi==0.104.1
uvicorn==0.24.0
PyJWT==2.8.0

# Research Agent
google-generativeai>=0.3.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0

# Utilities
python-dateutil==2.9.0.post0
regex==2024.9.11

